api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "amqp==5.3.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appdirs==1.4.4",
    "asgiref==3.8.1",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "auth0-python==4.9.0",
    "autobahn==24.4.2",
    "automat==25.4.16",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "billiard==4.2.1",
    "blinker==1.9.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "celery==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "channels==4.2.2",
    "channels-redis==4.2.1",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "coloredlogs==15.0.1",
    "constantly==23.10.4",
    "cryptography==45.0.2",
    "daphne==4.2.0",
    "decorator==5.2.1",
    "deprecated==1.2.18",
    "distro==1.9.0",
    "dj-database-url>=2.3.0",
    "django==4.2.21",
    "django-cors-headers==4.7.0",
    "djangorestframework==3.16.0",
    "docstring-parser==0.16",
    "durationpy==0.10",
    "et-xmlfile==2.0.0",
    "executing==2.2.0",
    "fastapi==0.115.9",
    "filelock==3.18.0",
    "flatbuffers==25.2.10",
    "frozenlist==1.6.0",
    "fsspec==2025.5.0",
    "google-auth==2.40.2",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.71.0",
    "h11==0.16.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "huggingface-hub==0.31.4",
    "humanfriendly==10.0",
    "hyperframe==6.1.0",
    "hyperlink==21.0.0",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "importlib-resources==6.5.2",
    "incremental==24.7.2",
    "instructor==1.8.3",
    "ipython==9.2.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.8.2",
    "json-repair==0.46.0",
    "json5==0.12.0",
    "jsonpickle==4.1.0",
    "jsonref==1.1.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "kombu==5.5.3",
    "kubernetes==32.0.1",
    "litellm==1.68.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mistralai>=1.7.1",
    "mmh3==5.1.0",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "multidict==6.4.4",
    "networkx==3.4.2",
    "numpy==2.2.6",
    "oauthlib==3.2.2",
    "openai==1.75.0",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.33.1",
    "opentelemetry-exporter-otlp-proto-common==1.33.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.33.1",
    "opentelemetry-exporter-otlp-proto-http==1.33.1",
    "opentelemetry-instrumentation==0.54b1",
    "opentelemetry-instrumentation-asgi==0.54b1",
    "opentelemetry-instrumentation-fastapi==0.54b1",
    "opentelemetry-proto==1.33.1",
    "opentelemetry-sdk==1.33.1",
    "opentelemetry-semantic-conventions==0.54b1",
    "opentelemetry-util-http==0.54b1",
    "orjson==3.10.18",
    "overrides==7.7.0",
    "packaging==25.0",
    "parso==0.8.4",
    "pdfminer-six==20250327",
    "pdfplumber==0.11.6",
    "pexpect==4.9.0",
    "pgvector==0.4.1",
    "pillow==11.2.1",
    "posthog==4.1.0",
    "priority==1.3.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.1",
    "protobuf==5.29.4",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pyopenssl==25.1.0",
    "pypdfium2==4.30.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pyvis==0.3.2",
    "pyyaml==6.0.2",
    "redis==4.6.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rich==13.9.4",
    "rpds-py==0.25.1",
    "rsa==4.9.1",
    "service-identity==24.2.0",
    "setuptools==80.8.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "stack-data==0.6.3",
    "starlette==0.45.3",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "tomli==2.2.1",
    "tomli-w==1.2.0",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "twisted==24.11.0",
    "txaio==23.1.1",
    "typer==0.15.4",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "uv==0.7.7",
    "uvicorn==0.34.2",
    "uvloop==0.21.0",
    "vine==5.1.0",
    "watchfiles==1.0.5",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "whitenoise==6.9.0",
    "wrapt==1.17.2",
    "yarl==1.20.0",
    "zipp==3.21.0",
    "zope-interface==7.2",
]
